refactor(parse): split signal/parse.rs into 5-file submodule (closes #455)
2579-line parse.rs reorganised into src/signal/parse/ with five
sibling files plus mod.rs. Each submodule is under 600 LOC and has a
single concern; mod.rs holds the module declarations, the two public
re-exports, and the existing test block (1273 LOC of tests stayed
grouped near the public surface they exercise).

File breakdown:
  envelope.rs (241) - parse_signal_event, parse_receive_event,
                      typing/receipt/read-sync routing
  message.rs  (538) - parse_data_message, parse_sent_sync,
                      parse_common_message_fields, parse_data_payload_event,
                      reactions, edits
  poll.rs     (129) - poll create / vote / terminate
  rpc.rs      (164) - parse_rpc_result + listContacts/Groups
  helpers.rs  (286) - attachment, link previews, mentions, text styles,
                      expiration formatting, MIME mapping
  mod.rs     (1273) - module declarations, re-exports, full test block

Cross-module helpers (envelope_source, format_expiration,
parse_attachment, etc) are pub(super) so siblings can call them
without making them crate-visible. The public surface is unchanged:
parse_signal_event and parse_rpc_result remain the only exports.

Built on top of #448 (parse dedup); the CommonMessageFields struct
and parse_data_payload_event helper introduced there land in
message.rs as part of this split.

Zero behaviour change. 559/559 tests still pass.

Note for merge order: this PR depends on #448 being merged first.
Filed and tracked as the parse_data_message dedup creates the helpers
that this split places into message.rs.
